+++ test/help.spec.js
@@ -1,5 +1,5 @@
 const header = require('./db/header');
-const tools = require('./tools');
+const tools = require('./db/tools');
 
 const promise = header.defPromise;
 const options = {
