+++ test/parameterized.spec.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const header = require('./db/header');
-const tools = require('./tools');
+const tools = require('./db/tools');
 
 const promise = header.defPromise;
 const options = {
